overlays work again, plonetabs, tinymce, css.
Default resources go into the overlay as well, The toolbar initialisation code
is in a seperate bundle now, to allow it to be included only once, in the main
document. Events are triggered on the iframe element (toolbar.el) and those
allow you to hook other code that executes within the context of the iframe.
These hooks presently initialise tinymce and ploneTabs.