For 'Suite Setup' I decided to use a slightly modify version of plone.act's 'Start browser and wake Plone up' keyword.  This plone.act keyword does some strange things at the end starting with
    Wait until keyword succeeds  30s  1s  Log in as site owner
    Log out
    Zodb teardown

First I'm not sure if a ZSERVER_FIXTURE site has a site owner (which reading
from the plone.app.testing and plone.testing doc I would guess yes) or whether
that line has an issue because the 'Log in as site owner' keyword doesn't have
a return value so wait until keyword suceeds will never succeed.  I suspect
the latter. Also I am not sure what the last two lines (Log out,
Zodb teardown) are supposed to do in this context and am assuming they were
left in by accident in the development of the library. So, for now, I am
using my own version of this keyword.

Modify test slightly giving a better test name and created tests to check
for toolbar whether user logged in or not.  Realized I will need to make a
plone.app.toolbar log in keyword because the plone.act expects standard
Plone UI. More consideration should be taken with design of core library
and various UIs.